// src/components/Buttons/enum.ts
export const enum ButtonState {
  DEFAULT = 'default',
  LOADING = 'loading',
  PRESSED = 'pressed',
  DISABLED = 'disabled',
}

export enum ButtonSize {
  LARGE = 'large',
  MEDIUM = 'medium',
  SMALL = 'small',
}

export const enum ButtonTypes {
  PRIMARY = 'primary',
  SECONDARY = 'secondary',
  LINK = 'link',
}

export const enum IconPosition {
  CENTER = 'center',
  LEFT = 'left',
  RIGHT = 'right',
  CENTER_RIGHT = 'center-right',
}

export enum IconButtonSize {
  LARGE = 'large',
  SMALL = 'small',
  MEDIUM = 'medium',
}

export const enum IconButtonType {
  PRIMARY = 'primary',
  SECONDARY = 'secondary',
}